Simple unsigned/signed comparison fix for cityhash

+++ cityhash.cc
@@ -17,9 +17,9 @@ uint64_t ntime()
 
 int main(int argc, char **argv)
 {
-  int begin = strtol(argv[1], NULL, 0);
-  int end = strtol(argv[2], NULL, 0);
-  int inc = strtol(argv[3], NULL, 0);
+  unsigned int begin = strtol(argv[1], NULL, 0);
+  unsigned int end = strtol(argv[2], NULL, 0);
+  unsigned int inc = strtol(argv[3], NULL, 0);
   uint64_t i, t1, t2, h = 0, n = 10000000;
   float t;
   size_t len;
@@ -28,7 +28,7 @@ int main(int argc, char **argv)
   for (i = 0; i < end; i ++)
     in[i] = rand() % 255 + 1;
   
-  for (len = begin; len < end; len += inc)
+  for (len = (size_t)begin; len < end; len += inc)
     {
       t1 = ntime();
       for (i = 0; i < n; i ++)
